+++ CHANGELOG.rst
@@ -1,6 +1,8 @@
 Release 0.14.0 (unreleased)
 ===========================
 
+* Set SBOM ``licenses`` to ``NOASSERTION`` when a license file is not found or cannot be classified (#1112)
+* Add a ``dfetch:license:finding`` property to SBOM (#1112)
 * Use github purl, repo and version for a github release archive in SBOM (#1063)
 * Allow ``dfetch freeze`` to accept project names to freeze only specific projects (#1063)
 * Edit manifest in-place when freezing inside a git or SVN superproject, preserving comments and layout (#1063)